Define CRYPTO_once_t as LONG on Windows.
This is used with a platform API, so it should use the corresponding
platform type, saving us the size assert. It's ever defined in an
internal header, so we can freely use windows.h and friends.

Reviewed-on: https://boringssl-review.googlesource.com/4680

+++ crypto/internal.h
@@ -336,7 +336,7 @@ static inline int constant_time_select_int(unsigned int mask, int a, int b) {
 typedef pthread_once_t CRYPTO_once_t;
 #define CRYPTO_ONCE_INIT PTHREAD_ONCE_INIT
 #else
-typedef int32_t CRYPTO_once_t;
+typedef LONG CRYPTO_once_t;
 #define CRYPTO_ONCE_INIT 0
 #endif
